Strip HTML from the summary
The summary tree is quite fragile in regards to markup and styles. If
an item's name consists of a <ul> element, the summary will look wrong.
Simply remove all markup from the summary items; they are not need
anyway, as we tend to only show the first few words of items.